jwt user

+++ src/models/todoPgModel.js
@@ -2,24 +2,23 @@ export default class TodoPgModel {
   constructor(client) {
     this.client = client;
   }
-
-  async getAll() {
-    const res = await this.client.query('SELECT id, titre, fait FROM todos ORDER BY created_at');
+  async getAll(userId) {
+    const res = await this.client.query('SELECT id, titre, fait FROM todos WHERE owner = $1 ORDER BY created_at', [userId]);
     return res.rows;
   }
 
-  async addTask(titre) {
-    const res = await this.client.query('INSERT INTO todos (titre) VALUES ($1) RETURNING id, titre, fait', [titre]);
+  async addTask(titre, userId) {
+    const res = await this.client.query('INSERT INTO todos (titre, owner) VALUES ($1, $2) RETURNING id, titre, fait', [titre, userId]);
     return res.rows[0];
   }
 
-  async completeTask(id) {
-    const res = await this.client.query('UPDATE todos SET fait = true WHERE id = $1 RETURNING id', [id]);
+  async completeTask(id, userId) {
+    const res = await this.client.query('UPDATE todos SET fait = true WHERE id = $1 AND owner = $2 RETURNING id', [id, userId]);
     return res.rowCount > 0;
   }
 
-  async deleteTask(id) {
-    const res = await this.client.query('DELETE FROM todos WHERE id = $1', [id]);
+  async deleteTask(id, userId) {
+    const res = await this.client.query('DELETE FROM todos WHERE id = $1 AND owner = $2', [id, userId]);
     return res.rowCount > 0;
   }
 }
